Refector: utils폴더에 함수 모듈화
