+++ signin.html
@@ -16,7 +16,7 @@
         </div>
         <p class="question">회원이 아니신가요? <a href="/signup">회원 가입하기</a></p>
       </section>
-      <form class="middle" name="formname">
+      <form class="middle" name="formname" action="./folder">
         <div class="input">
           <p>이메일</p>
           <input id="email" type="email" placeholder="codeit@codeit.com" />
@@ -39,6 +39,6 @@
         </div>
       </section>
     </main>
-    <script src="signin.js"></script>
+    <script src="signin.js" type="module"></script>
   </body>
 </html>
